+++ src/Components/Socials.php
@@ -11,9 +11,20 @@ enum Service: string
 {
     case WEBSITE = 'WEBSITE';
     case EMAIL = 'EMAIL';
+    case PHONE = 'PHONE';
+    case FACEBOOK = 'FACEBOOK';
+    case YOUTUBE = 'YOUTUBE';
     case TWITTER = 'TWITTER';
     case X = 'X';
     case GITHUB = 'GITHUB';
+    case INSTAGRAM = 'INSTAGRAM';
+    case LINKEDIN = 'LINKEDIN';
+    case SLACK = 'SLACK';
+    case DISCORD = 'DISCORD';
+    case TIKTOK = 'TIKTOK';
+    case SNAPCHAT = 'SNAPCHAT';
+    case THREADS = 'THREADS';
+    case TELEGRAM = 'TELEGRAM';
 }
 
 class Item implements JsonSerializable