+++ CHANGELOG.md
@@ -1,7 +1,19 @@
 # CHANGELOG
 
 **Unpublished:**
-- Introduced CHANGELOG
+- Introduced `CHANGELOG.md`
+- Introduced new services as social icons:
+  - `Service::PHONE` (converts into a link with `tel:` prefix)
+  - `Service::FACEBOOK`
+  - `Service::YOUTUBE`
+  - `Service::INSTAGRAM`
+  - `Service::LINKEDIN`
+  - `Service::SLACK`
+  - `Service::DISCORD`
+  - `Service::TIKTOK`
+  - `Service::SNAPCHAT`
+  - `Service::THREADS`
+  - `Service::TELEGRAM`
 
 **v0.1.0-alpha.0:**
 - Initial implementation